requirements.txt
tensorflow
numpy
Pillow
mysqlclient
matplotlib
seaborn
scikit-learn
opencv-python
keras
tensorflow.keras
